requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
databench-eval==3.0.1
datasets==3.1.0
debugpy==1.8.12
decorator==5.2.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jiter==0.8.2
jupyter-client==8.6.3
jupyter-core==5.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==1.24.4
openai==1.65.1
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.50
propcache==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2